+++ packages/client/src/app.tsx
@@ -1,48 +1,13 @@
-import { useState, type ChangeEvent } from "react";
-import SearchBar from "./components/SearchBar/SearchBar";
-import ResultsList from "./components/ResultsList/ResultsList";
-import useSearchData from "./hooks/useSearchData";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import SearchPage from "./pages/SearchPage";
 
 function App() {
-  const [searchValue, setSearchValue] = useState("");
-  const [showClearBtn, setShowClearBtn] = useState(false);
-  // Use the hook
-  const { hotels, countries, cities } = useSearchData(searchValue);
-
-  const fetchData = (event: ChangeEvent<HTMLInputElement>) => {
-    const value = event.target.value;
-    setSearchValue(value);
-    setShowClearBtn(!!value);
-  };
-
-  const clearSearch = () => {
-    setSearchValue("");
-    setShowClearBtn(false);
-  };
-
   return (
-    <div className="App">
-      <div className="container">
-        <div className="row height d-flex justify-content-center align-items-center">
-          <div className="col-md-6">
-            <div className="dropdown">
-              <SearchBar
-                value={searchValue}
-                onChange={fetchData}
-                onClear={clearSearch}
-                showClearBtn={showClearBtn}
-              />
-              <ResultsList
-                hotels={hotels}
-                countries={countries}
-                cities={cities}
-                showResults={!!searchValue}
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/" element={<SearchPage />} />
+      </Routes>
+    </Router>
   );
 }
